Use notify/readOnly for results
This changes the name of the event, and is more semantically correct. `searchResults` should not be modifiable except internally, so we can leverage Polymer for this.

Another slight change here is that hitting the X (or clearing the field and hitting Enter) will send an empty `search-field-name` value. For most reasonable APIs, this should give you all results so it goes back to being "unfiltered".