Test feedback
Icon was not properly changing state when the button was clicked (only working properly with Enter key). Note that this approach means using the .search() and .clear() API methods don't necessarily update the icons as one might expect... but it's better than it was. Might be a better solution, though.