Break out style and behavior
In an attempt to make this a bit more modular, this just breaks out the behavior and style into their own files. The real intent behind this is that the customized search widget in My Courses should be able to mostly use the existing style and behavior here, and only override things as required.