US90475 Remove built-in caching from d2l-search-widget
Rather than implementing this here, we should just recommend that the consumer of the element also use the d2l-fetch-simple-cache middleware, along with d2l-fetch-auth. This allows the component to be much simpler, and prevents us from having a double-caching situation in the (likely) case where the user _is_ using d2l-fetch-simple-cache.

This is a breaking change, however, so this will be released with a major version bump.