US97068 Set searchInput to query value
If we pass in a searchAction that has a search/query field with a non-empty value, we should be setting the seach input value to that value. This is "more correct" behaviour than previously existed, as if we were sending in a search action that already _had_ a query value, the presumption is that we're working with a collection that has already been searched - therefore, we should reflect this state in the search input field.